# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server for Suasor",
        "title": "Suasor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dane Miller",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticates a user with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in a user",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated user",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_AuthDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or inactive account",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Invalidates the refresh token, effectively logging the user out",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log out a user",
                "parameters": [
                    {
                        "description": "Logout data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-any"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Invalid refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Get a new access token using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully refreshed token",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_AuthDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Registers a new user with email, username, and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered user",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_AuthDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "409": {
                        "description": "Email or username already in use",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/auth/validate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates the user's session token and returns current user profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validate user session",
                "responses": {
                    "200": {
                        "description": "Valid session with user details",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_UserResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired session token",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.ErrorType": {
            "type": "string",
            "enum": [
                "FAILED_CHECK",
                "UNAUTHORIZED",
                "NOT_FOUND",
                "BAD_REQUEST",
                "INTERNAL_ERROR",
                "FORBIDDEN",
                "CONFLICT",
                "VALIDATION_ERROR",
                "RATE_LIMITED",
                "TIMEOUT",
                "SERVICE_UNAVAILABLE",
                "UNPROCESSABLE_ENTITY"
            ],
            "x-enum-varnames": [
                "ErrorTypeFailedCheck",
                "ErrorTypeUnauthorized",
                "ErrorTypeNotFound",
                "ErrorTypeBadRequest",
                "ErrorTypeInternalError",
                "ErrorTypeForbidden",
                "ErrorTypeConflict",
                "ErrorTypeValidation",
                "ErrorTypeRateLimited",
                "ErrorTypeTimeout",
                "ErrorTypeServiceUnavailable",
                "ErrorTypeUnprocessableEntity"
            ]
        },
...
        "requests.LoginRequest": {
            "description": "Request payload for user login",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email is the user's email address\n@Description User's email address\n@Example \"user@example.com\"",
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "description": "Password is the user's chosen password\n@Description User's password (plain text in request)",
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "requests.LogoutRequest": {
            "description": "Request payload for user logout",
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "description": "RefreshToken identifies the session to terminate\n@Description JWT refresh token to invalidate",
                    "type": "string"
                }
            }
        },
...
        "requests.RefreshTokenRequest": {
            "description": "Request payload for refreshing access tokens",
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "description": "RefreshToken is used to generate a new access token\n@Description JWT refresh token to use for generating new access token",
                    "type": "string"
                }
            }
        },
        "requests.RegisterRequest": {
            "description": "Request payload for user registration",
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "Email is the user's email address\n@Description User's email address\n@Example \"user@example.com\"",
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "description": "Password is the user's chosen password\n@Description User's password (plain text in request)",
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "description": "Username is the user's desired username\n@Description User's chosen username\n@Example \"johndoe\"",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "responses.APIResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.APIResponse-responses_AuthDataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.AuthDataResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.APIResponse-responses_UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.UserResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.AuthDataResponse": {
            "description": "Authentication data returned to client after successful authentication",
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "AccessToken is the JWT token for API access\n@Description JWT access token for authenticated requests\n@Example \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
                    "type": "string"
                },
                "expiresAt": {
                    "description": "ExpiresAt indicates when the access token expires\n@Description UNIX timestamp when the access token expires\n@Example 1674140400",
                    "type": "integer"
                },
                "refreshToken": {
                    "description": "RefreshToken is used to get new access tokens\n@Description JWT refresh token for obtaining new access tokens\n@Example \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
                    "type": "string"
                },
                "user": {
                    "description": "User contains the user profile information\n@Description User profile data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    ]
                }
            }
        },
...
        "responses.ErrorDetails": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "stackInfo": {
                    "type": "string"
                }
            }
        },
...
        "responses.ErrorResponse-responses_ErrorDetails": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/responses.ErrorDetails"
                },
                "message": {
                    "type": "string",
                    "example": "This is a pretty message"
                },
                "request_id": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 201
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorType"
                        }
                    ],
                    "example": "FAILED_CHECK"
                }
            }
        },
...
        "responses.UserResponse": {
            "description": "User information returned in API responses",
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email is the unique email address of the user\n@Description User's email address\n@Example \"user@example.com\"",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier for the user\n@Description User's unique identifier\n@Example 1",
                    "type": "integer"
                },
                "role": {
                    "description": "Role defines the user's permission level\n@Description User's role in the system\n@Enum \"user\" \"admin\"\n@Example \"user\"",
                    "type": "string"
                },
                "username": {
                    "description": "Username is the display name chosen by the user\n@Description User's chosen username\n@Example \"johndoe\"",
                    "type": "string"
                }
            }
        },
...
    }
}